+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Master (Unreleased)
 
+### Fixed
+
+* Unfriendly output in byebug's executable when no script specified (#256).
+* Unfriendly output in byebug's executable when script doesn't exist.
+* Unfriendly output in byebug's executable when script has invalid code.
+
 ## 9.0.2 - 2016-05-15
 
 ### Fixed