+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 
 ## 9.0.1 - 2016-05-14
 
+### Fixed
+
 * `quit` never exiting when remote debugging (#201).
 
 ## 9.0.0 - 2016-05-11